refactor: AfternoteTypography 위치 이동 및 관련 파일 정리
- `AfternoteTypography` 데이터 클래스를 `Theme.kt`에서 `Type.kt`로 이동하여 폰트 정의와 타이포그래피 설정을 한 곳에서 관리하도록 개선
- `Type.kt`에 필요한 임포트(`TextStyle`, `sp`, `em` 등) 추가 및 구조 정리
